Harden plan tool clear

+++ internal/agent/plan_tool.go
@@ -428,6 +428,19 @@ func writePlanState(path string, st planState) error {
 }
 
 func clearPlanState(path string) error {
+	info, err := os.Lstat(path)
+	if errors.Is(err, os.ErrNotExist) {
+		return nil
+	}
+	if err != nil {
+		return err
+	}
+	if info.IsDir() {
+		return errors.New("plan path is a directory")
+	}
+	if info.Mode()&os.ModeSymlink != 0 {
+		return errors.New("plan path must not be a symlink")
+	}
 	if err := os.Remove(path); err != nil && !errors.Is(err, os.ErrNotExist) {
 		return err
 	}
